refactor: controller 파일 정리
- productAddMenu, productPurchaseMenu, vendingMachineChargeMenu, localStorage 폴더로 각각 기능을 담당하는 파일 이동
