+++ src/controller/ProductPurchaseMenu/CheckProductPurchaseMenu.js
@@ -1,4 +1,4 @@
-import { DOM, LOCAL_STORAGE, EVENT, TEMPLATE } from '../utils/constant.js';
+import { DOM, LOCAL_STORAGE, EVENT, TEMPLATE } from '../../utils/constant.js';
 import CheckCoinReturn from './CheckCoinReturn.js';
 import SetVendingMachinePurchase from './SetVendingMachinePurchase.js';
